agent_server/requirements.txt
fastapi
uvicorn[standard]
openai
openai-agents
openai-agents-mcp
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2 
supabase
